requirements.txt
aiofiles==25.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
curl_cffi==0.13.0
dataclasses-json==0.6.7
emoji==2.15.0
filetype==1.2.0
frozendict==2.4.7
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.0.7
langchain-core==1.0.5
langchain-text-splitters==1.0.0
langdetect==1.0.9
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.43
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.1
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.11.0
nltk==3.9.2
numpy==2.3.4
olefile==0.47
ollama==0.6.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
protobuf==6.33.1
psutil==7.1.3
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pymilvus==2.6.3
pypdf==6.3.0
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-iso639==2025.11.11
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.29.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
streamlit==1.51.0
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.18.20
unstructured-client==0.42.4
urllib3==2.5.0
webencodings==0.5.1
websockets==15.0.1
wrapt==2.0.1
xxhash==3.6.0
yfinance==0.2.66
zstandard==0.25.0
